Source update

+++ src/memory/wii/mem2_manager.c
@@ -3,6 +3,7 @@
 #include <unistd.h>
 
 #include <ogc/machine/asm.h>
+#include <ogc/machine/processor.h>
 #include <ogc/lwp_heap.h>
 #include <ogc/system.h>
 
@@ -163,9 +164,8 @@ char *_mem2_strdup(const char *s)
 
     if (s)
     {
-        int len = strlen(s) + 1;
-
-        ptr = _mem2_calloc(1, len);
+        size_t len = strlen(s) + 1;
+        ptr        = _mem2_calloc(1, len);
 
         if (ptr)
             memcpy(ptr, s, len);
